refactor: separate validation in the utils
